+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to `laravel-webhook-server` will be documented in this file
 
+## 1.11.3 - 2021-04-02
+
+- fix for missing default headers when using withHeaders (#79)
+
+## 1.11.2 - 2021-03-17
+
+- dispatch should return the PendingDispatch (#74)
+
+## 1.11.1 - 2020-12-15
+
+- fix exception name for invalid signers (#67)
+
 ## 1.11.0 - 2020-11-28
 
 - add support for PHP 8